Move functions to miscFunctions.py
loadPickle
selectMatrices
loadNPZ
